`get<N>()` for reading sequence of bytes
+ This is a proper solution for reading a sequence of bytes instead of an approach like:

```
uint8_t bytes[48];
for(uint8_t &c : bytes)
  c = uart.get();
```

The above method is not appropriate to receive a sequence of bytes in
a row because some CPU cycles (from application code) will be consumed
to handle each received byte. The sender can start transmitting the
next byte before the receiver is waiting for its start bit. The sender
can be faster than the receiver, outside the time-sensitive zone. In
other words, the synchronization between the transmitter and receiver
occurs for each byte, and between two bytes, the sender can begin
sending before the receiver waits for the next start bit.