Offer a handshaking approach
+ The methods `request_to_send()` and `clear_to_send()` were added as
an optional handshaking approach to ensure that the receiver is ready
to receive data from the transmitter. The support for calling `get()`
and `get_bytes()` while informing the number of CPU cycles that have
already been executed was removed. I don't think it's a good approach
to implement an async solution. It is highly dependent on the
generated code and can be fragile, as the compiler can easily change
the number of instructions executed before reaching
`get()`. Additionally, when a byte is ready, the event loop may take
too much time to reach the point of consuming the byte.